Add integration mode check in CreateAWSSSOProfile function
This commit introduces a check for the AWS integration mode within the CreateAWSSSOProfile function. If the integration mode is set to "native," the function will return early without creating a profile. The default mode is set to "cone" if not specified, enhancing the flexibility and configurability of AWS SSO profile creation in the Cone CLI.